requirements.txt
flask==3.1.2
mysqlclient==2.2.7
python-dotenv==1.1.1
SQLAlchemy==2.0.44
pydantic==2.12.3
pyJWT==2.10.1
email-validator==2.2.0
psycopg2-binary==2.9.11
firebase-admin==7.1.0